refactor: move view's package
