+++ src/main/kotlin/racingcar/Main.kt
@@ -1,7 +1,5 @@
 package racingcar
 
-import racingcar.view.View
-
 fun main() {
     val view = View()
     view.playRacingGame()
